Created File Model, Added file upload.
When an User drags or selects a file in the dropzone, a file will be
uploaded to review.File and binded to the review.Analysis and it's
current state.
The template also displays a list of files both external_reports and
reviews. All of them can be downloaded from the same modal.